Softomation/HighwaySolutions/WindowsService/PyLaneService/requirements.txt
appdirs==1.4.4
argcomplete==3.5.3
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
cryptography==44.0.0
Cython==3.0.11
Flask==3.0.3
Flask-Cors==4.0.0
gmssl==3.2.2
hidapi==0.14.0
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector-python==8.3.0
netifaces==0.11.0
numpy==1.26.4
opencv-python==4.5.5.64
paho-mqtt==1.6.1
pandas==2.2.2
paramiko==3.4.0
pillow==11.1.0
ping3==4.0.8
psutil==2.1.1
pycparser==2.22
pycryptodomex==3.21.0
PyJWT==2.8.0
PyNaCl==1.5.0
Pypubsub==4.0.3
pyserial==3.5
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-escpos==3.1
pytz==2024.2
pyusb==1.2.1
PyYAML==6.0.2
qrcode==8.0
requests==2.32.3
RFIDReaderAPI @ file:///D:/GitProjects/HighwaySolutions/Softomation/HighwaySolutions/WindowsService/PyLaneService/libs/RFIDReaderAPI-1.1-py3-none-any.whl#sha256=b407f38187825e3d93e5da5d21cf40b92a7222e33376de6d84f16033887c68e9
scapy==2.5.0
six==1.17.0
tqdm==4.67.1
tzdata==2024.2
urllib3==2.3.0
vidgear==0.3.2
websockets==12.0
Werkzeug==3.1.3
